Move functions that are only used in grid.py to grid.py.
