+++ harmony_regridding_service/grid.py
@@ -7,20 +7,18 @@ import xarray as xr
 from harmony_service_lib.message import Message as HarmonyMessage
 from harmony_service_lib.message_utility import has_dimensions
 from netCDF4 import Dataset
+from pyproj import CRS
+from pyproj.exceptions import CRSError
 from pyresample import create_area_def
 from pyresample.geometry import AreaDefinition, SwathDefinition
 from varinfo import VarInfoFromNetCDF4
 
-from harmony_regridding_service.crs import (
-    crs_from_source_data,
-)
 from harmony_regridding_service.dimensions import (
-    dims_are_lon_lat,
-    dims_are_projected_x_y,
     get_column_dims,
     get_row_dims,
 )
 from harmony_regridding_service.exceptions import (
+    InvalidSourceCRS,
     InvalidSourceDimensions,
     SourceDataError,
 )
@@ -241,3 +239,55 @@ def compute_array_bounds(values: np.ndarray) -> tuple[np.float64, np.float64]:
     left = values[0] - half_width
     right = values[-1] + half_width
     return (left, right)
+
+
+def crs_from_source_data(dt: xr.DataTree, variables: set) -> CRS:
+    """Create a CRS describing the grid in the source file.
+
+    Look through the variables for metadata that points to a grid_mapping
+    and generate a CRS from that information.
+
+    The metadata is not always clear or easy to parse into a CRS. Take a
+    shortcut when possible.
+
+    if the grid_mapping has a known EASE2 grid name, use the EPSG code known
+    apriori.
+
+    Args:
+      dt: the source file as an opened DataTree
+
+      variables: set of variables all sharing the same 2-dimensional grid is
+                 traversed looking for a grid_mapping.
+
+    Returns:
+      CRS object
+
+    """
+    for varname in variables:
+        var = dt[varname]
+        if 'grid_mapping' in var.attrs:
+            try:
+                return CRS.from_cf(dt[var.attrs['grid_mapping']].attrs)
+            except CRSError as e:
+                raise InvalidSourceCRS(
+                    'Could not create a CRS from grid_mapping metadata'
+                ) from e
+
+    raise InvalidSourceCRS('No grid_mapping metadata found.')
+
+
+def dims_are_lon_lat(dimensions: tuple[str, str], var_info: VarInfoFromNetCDF4) -> bool:
+    """Does the dimension pair represent longitudes/latitudes."""
+    return all(
+        var_info.get_variable(dim_name).is_geographic() for dim_name in dimensions
+    )
+
+
+def dims_are_projected_x_y(
+    dimensions: tuple[str, str], var_info: VarInfoFromNetCDF4
+) -> bool:
+    """Does the dimension pair represent projected x/y values."""
+    return all(
+        var_info.get_variable(dim_name).is_projection_x_or_y()
+        for dim_name in dimensions
+    )
